WIP test: contact and metadata db integration test
before adding this functionality to the main function, I have to make sure it works in a test env

I already found a bug, plus, it looks like I'll need to do some Rust magic because I can't have 2 in-memory connections at the same time.
I wanted to do this anyway; as I believe it'll be more efficient.